feat: :tada: Title: Initialize AuthV5-toolkit project
Details:
This is the initial commit for the AuthV5-toolkit project.

The AuthV5-toolkit is a repository for a toolkit designed to facilitate the building of advanced custom authentication into next.js applications.  In this commit, we are setting up the basic project structure and adding the initial files necessary to start the development. This includes the layout.tsx file, which contains the project description.  As we move forward, we will be adding more functionality and features to this toolkit. This initial commit serves as the foundation for future development and improvements.

Remember, this is just the beginning of our journey towards creating a robust and efficient authentication toolkit for next.js applications.