docs: :memo: docs(readme): Expand project description in README.md
Description:
This commit expands the project description in the README.md file to include information about the client-side and server-side aspects and approaches demonstrated by the Auth.js V5 Toolkit.
- In `README.md`, updates have been made to handle the expansion of the project description.

Changes:
- Update `README.md` to include additional information about the client-side and server-side aspects and approaches demonstrated by the Auth.js V5 Toolkit.

This commit doesn't introduce any breaking changes. The updates are compatible with the existing codebase and provide a more comprehensive overview of the project for new users or contributors who are reading the README.md file.
Co-authored-by: Ricardo Esteves <30535937+RicardoGEsteves@users.noreply.github.com>