docs: :memo: docs(update-readme): Update Concept section in README.md
Description:
This commit introduces an update to the "Concept" section in the README.md file. The changes are part of the ongoing work on the `update-readme` branch.
- In `README.md`, the "Concept" section has been updated to include a new subheading for "Client-Side Implementation".

Changes:
- Update `README.md` to include a new subheading for "Client-Side Implementation" in the "Concept" section

This commit doesn't introduce any breaking changes. The updates are compatible with the existing codebase and provide a more comprehensive and structured overview of the project concept for new users or contributors who are reading the README.md file.
Co-authored-by: Ricardo Esteves <30535937+RicardoGEsteves@users.noreply.github.com>